# market_dataset_service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinSight Backtesting API",
    "description": "\n    Advanced backtesting system for cryptocurrency trading strategies.\n    \n    ## Features\n    \n    * **Administrative Operations**: Data management, system statistics, maintenance\n    * **Market Data Management**: Fetch, store, and convert OHLCV data across timeframes\n    * **Strategy Backtesting**: Run and analyze trading strategies with multiple engines\n    * **Cross-Repository Support**: MongoDB, CSV, InfluxDB storage adapters\n    \n    ## Authentication\n    \n    Admin endpoints require API key authentication via Bearer token.\n    Set the `ADMIN_API_KEY` environment variable or use the default key.\n    \n    ## Architecture\n    \n    Built with Ports & Adapters (Hexagonal Architecture) for maximum flexibility:\n    - **Service Layer**: Business logic and orchestration\n    - **Repository Layer**: Pluggable storage adapters  \n    - **Strategy Layer**: Extensible trading strategies\n    - **Adapter Layer**: Exchange connectors and backtesting engines\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/backtesting/strategies": {
      "get": {
        "tags": ["backtesting", "backtesting"],
        "summary": "Get Available Strategies",
        "description": "Get list of available trading strategies.\n\nReturns all registered strategies with their descriptions,\nparameters, and configuration options.",
        "operationId": "get_available_strategies_api_v1_backtesting_strategies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters"
          },
          "404": {
            "description": "Not found - Strategy or data not available"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/backtesting/strategies/{strategy_name}/config": {
      "get": {
        "tags": ["backtesting", "backtesting"],
        "summary": "Get Strategy Config",
        "description": "Get configuration schema for a specific strategy.\n\nReturns the parameter definitions, default values,\nand validation rules for the strategy.",
        "operationId": "get_strategy_config_api_v1_backtesting_strategies__strategy_name__config_get",
        "parameters": [
          {
            "name": "strategy_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyConfigSchema"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters"
          },
          "404": {
            "description": "Not found - Strategy or data not available"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtesting/run": {
      "post": {
        "tags": ["backtesting", "backtesting"],
        "summary": "Run Backtest",
        "description": "Execute a backtest with specified parameters.\n\nRuns the specified strategy against historical market data\nand returns comprehensive performance metrics and analysis.",
        "operationId": "run_backtest_api_v1_backtesting_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters"
          },
          "404": {
            "description": "Not found - Strategy or data not available"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtesting/results/{backtest_id}": {
      "get": {
        "tags": ["backtesting", "backtesting"],
        "summary": "Get Backtest Results",
        "description": "Retrieve results from a previous backtest.\n\nReturns stored backtest results by ID, with options to include\nor exclude detailed trade and equity curve data.",
        "operationId": "get_backtest_results_api_v1_backtesting_results__backtest_id__get",
        "parameters": [
          {
            "name": "backtest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backtest Id"
            }
          },
          {
            "name": "include_trades",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include trade details",
              "default": true,
              "title": "Include Trades"
            },
            "description": "Include trade details"
          },
          {
            "name": "include_equity_curve",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include equity curve data",
              "default": true,
              "title": "Include Equity Curve"
            },
            "description": "Include equity curve data"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters"
          },
          "404": {
            "description": "Not found - Strategy or data not available"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["backtesting", "backtesting"],
        "summary": "Delete Backtest Results",
        "description": "Delete stored backtest results.\n\nRemoves backtest results and associated data from storage.\nThis operation cannot be undone.",
        "operationId": "delete_backtest_results_api_v1_backtesting_results__backtest_id__delete",
        "parameters": [
          {
            "name": "backtest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backtest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestDeletionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters"
          },
          "404": {
            "description": "Not found - Strategy or data not available"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtesting/history": {
      "get": {
        "tags": ["backtesting", "backtesting"],
        "summary": "Get Backtest History",
        "description": "Get history of recent backtests.\n\nReturns a paginated list of recent backtest executions with basic\nmetadata and performance summaries.",
        "operationId": "get_backtest_history_api_v1_backtesting_history_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "strategy_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by strategy name",
              "title": "Strategy Filter"
            },
            "description": "Filter by strategy name"
          },
          {
            "name": "symbol_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by symbol",
              "title": "Symbol Filter"
            },
            "description": "Filter by symbol"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters"
          },
          "404": {
            "description": "Not found - Strategy or data not available"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtesting/engines": {
      "get": {
        "tags": ["backtesting", "backtesting"],
        "summary": "Get Available Engines",
        "description": "Get list of available backtesting engines.\n\nReturns information about supported backtesting engines\nand their capabilities.",
        "operationId": "get_available_engines_api_v1_backtesting_engines_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestEnginesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters"
          },
          "404": {
            "description": "Not found - Strategy or data not available"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/backtesting/stats": {
      "get": {
        "tags": ["backtesting", "backtesting"],
        "summary": "Get Backtesting Stats",
        "description": "Get backtesting system statistics.\n\nReturns statistics about stored backtests, performance metrics,\nand system utilization.",
        "operationId": "get_backtesting_stats_api_v1_backtesting_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Backtesting Stats Api V1 Backtesting Stats Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters"
          },
          "404": {
            "description": "Not found - Strategy or data not available"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BacktestDeletionResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether deletion was successful"
          },
          "backtest_id": {
            "type": "string",
            "title": "Backtest Id",
            "description": "Deleted backtest identifier"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Status message"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": ["success", "backtest_id", "message"],
        "title": "BacktestDeletionResponse",
        "description": "Response schema for backtest deletion endpoint."
      },
...
      "BacktestRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "description": "Trading symbol (e.g., BTCUSDT)"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe",
            "description": "Timeframe (e.g., 1h, 1d)"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange",
            "description": "Exchange name",
            "default": "binance"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date",
            "description": "Backtest start date"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "title": "End Date",
            "description": "Backtest end date"
          },
          "strategy_type": {
            "$ref": "#/components/schemas/StrategyType",
            "description": "Strategy type to use"
          },
          "strategy_params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Strategy Params",
            "description": "Strategy-specific parameters"
          },
          "initial_capital": {
            "type": "number",
            "exclusiveMinimum": 0,
            "title": "Initial Capital",
            "description": "Initial capital for backtesting",
            "default": 10000
          },
          "commission": {
            "type": "number",
            "maximum": 1,
            "minimum": 0,
            "title": "Commission",
            "description": "Commission rate (0.001 = 0.1%)",
            "default": 0.001
          },
          "max_position_size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Position Size",
            "description": "Maximum position size as fraction of capital"
          },
          "stop_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss",
            "description": "Stop loss percentage (0.05 = 5%)"
          },
          "take_profit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit",
            "description": "Take profit percentage (0.10 = 10%)"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "timeframe",
          "start_date",
          "end_date",
          "strategy_type"
        ],
        "title": "BacktestRequest",
        "description": "Request schema for backtesting operations."
      },
      "BacktestResult": {
        "properties": {
          "backtest_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backtest Id",
            "description": "Unique backtest identifier"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "strategy_type": {
            "$ref": "#/components/schemas/StrategyType"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "title": "End Date"
          },
          "duration_days": {
            "type": "integer",
            "title": "Duration Days"
          },
          "initial_capital": {
            "type": "number",
            "title": "Initial Capital"
          },
          "final_capital": {
            "type": "number",
            "title": "Final Capital"
          },
          "metrics": {
            "$ref": "#/components/schemas/PerformanceMetrics"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/TradeResult"
            },
            "type": "array",
            "title": "Trades"
          },
          "equity_curve": {
            "items": {
              "$ref": "#/components/schemas/EquityCurvePoint"
            },
            "type": "array",
            "title": "Equity Curve"
          },
          "strategy_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Strategy Data",
            "description": "Strategy-specific output data"
          },
          "execution_time_seconds": {
            "type": "number",
            "title": "Execution Time Seconds"
          },
          "engine_used": {
            "type": "string",
            "title": "Engine Used"
          },
          "is_valid": {
            "type": "boolean",
            "title": "Is Valid",
            "description": "Whether results are valid",
            "default": true
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "Execution warnings"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "description": "Execution errors"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "timeframe",
          "exchange",
          "strategy_type",
          "start_date",
          "end_date",
          "duration_days",
          "initial_capital",
          "final_capital",
          "metrics",
          "trades",
          "equity_curve",
          "execution_time_seconds",
          "engine_used"
        ],
        "title": "BacktestResult",
        "description": "Comprehensive backtesting result."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StrategyConfigSchema": {
        "properties": {
          "strategy_name": {
            "type": "string",
            "title": "Strategy Name",
            "description": "Strategy name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Strategy description"
          },
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/ParameterSchema"
            },
            "type": "array",
            "title": "Parameters",
            "description": "Strategy parameters"
          },
          "examples": {
            "additionalProperties": true,
            "type": "object",
            "title": "Examples",
            "description": "Example configurations"
          }
        },
        "type": "object",
        "required": ["strategy_name", "description", "parameters"],
        "title": "StrategyConfigSchema",
        "description": "Response schema for strategy configuration endpoint."
      },
      "StrategyListResponse": {
        "properties": {
          "strategies": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Strategies",
            "description": "List of available strategies"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Total number of strategies"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": ["strategies", "count"],
        "title": "StrategyListResponse",
        "description": "Response schema for strategy list endpoint."
      },
      "StrategyType": {
        "type": "string",
        "enum": [
          "moving_average_crossover",
          "rsi_strategy",
          "bollinger_bands",
          "macd_strategy",
          "simple_buy_hold"
        ],
        "title": "StrategyType",
        "description": "Supported strategy types."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}